+++ modules/tinymce/src/themes/silver/main/ts/ui/core/FormatControls.ts
@@ -9,10 +9,10 @@ import Editor from 'tinymce/core/api/Editor';
 
 import { UiFactoryBackstage } from '../../backstage/Backstage';
 import * as AlignmentButtons from './AlignmentButtons';
+import * as ChoiceControls from './ChoiceControls';
 import * as ColorSwatch from './color/ColorSwatch';
 import * as ComplexControls from './ComplexControls';
 import * as IndentOutdent from './IndentOutdent';
-import * as LineHeight from './LineHeight';
 import * as SimpleControls from './SimpleControls';
 import * as UndoRedo from './UndoRedo';
 import * as VisualAid from './VisualAid';
@@ -25,7 +25,7 @@ const setup = (editor: Editor, backstage: UiFactoryBackstage) => {
   ColorSwatch.register(editor);
   VisualAid.register(editor);
   IndentOutdent.register(editor);
-  LineHeight.register(editor);
+  ChoiceControls.register(editor);
 };
 
 export {